+++ packages/expo-apple-authentication/package.json
@@ -2,8 +2,8 @@
   "name": "expo-apple-authentication",
   "version": "1.0.0",
   "description": "A unimodule that provides SignIn With Apple",
-  "main": "build/AppleAuthentication.js",
-  "types": "build/AppleAuthentication.d.ts",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
   "sideEffects": false,
   "scripts": {
     "build": "expo-module build",
