--- packages/expo-apple-authentication/src/ExpoAppleAuthenticationNative.ts
@@ -1,7 +1,0 @@
-import { NativeModulesProxy } from '@unimodules/core';
-
-export default NativeModulesProxy.ExpoAppleAuthentication || {
-  isAvailableAsync(): Promise<boolean> {
-    return Promise.resolve(false);
-  },
-} as any;
