+++ packages/expo-apple-authentication/src/index.ts
@@ -1,0 +1,3 @@
+export * from './AppleAuthentication';
+export * from './AppleAuthentication.types';
+export { default as SignInWithAppleButton } from './SignInWithAppleButton';
